+++ xen/include/asm-x86/mach-generic/mach_apic.h
@@ -21,9 +21,9 @@ static inline void enable_apic_mode(void)
 	/*es7000_sw_apic();*/
 }
 
-#define apicid_to_node(apicid) ((int)apicid_to_node[(u8)apicid])
+#define apicid_to_node(apicid) ((int)apicid_to_node[(u32)apicid])
 
-extern u8 bios_cpu_apicid[];
+extern u32 bios_cpu_apicid[];
 static inline int cpu_present_to_apicid(int mps_cpu)
 {
 	if (mps_cpu < NR_CPUS)
